requirements.txt
pymysql
djangorestframework
django-safedelete
jwt